Adding declaration of sparse CSRMatMulOp and CSRMatMulCPUOp in mat_mul_op.h for use by other kernels
